requirements.txt
django==3.0.3
autopep8==1.5
PyYAML==5.3
Pillow==7.0.0
mysqlclient==1.4.6
django-ckeditor==5.9.0
requests==  2.23.0
django-multiselectfield==0.1.12
django-redis==4.11.0
django-debug-toolbar==2.2
pusher==3.0.0
